requirements.txt
alembic==1.13.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==4.0.3 ; python_version >= "3.11" and python_version < "3.12.0"
asyncpg==0.29.0 ; python_version >= "3.11" and python_version < "4.0"
bcrypt==4.1.1 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.3.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.11" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
cryptography==41.0.7 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.5.0 ; python_version >= "3.11" and python_version < "4.0"
docker==7.0.0 ; python_version >= "3.11" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi[full]==0.104.1 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.16.2 ; python_version >= "3.11" and python_version < "4.0"
google-api-core[grpc]==2.16.2 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.27.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-firestore==2.14.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.62.0 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.0.2 ; platform_machine == "aarch64" and python_version >= "3.11" and python_version < "4.0" or platform_machine == "ppc64le" and python_version >= "3.11" and python_version < "4.0" or platform_machine == "x86_64" and python_version >= "3.11" and python_version < "4.0" or platform_machine == "amd64" and python_version >= "3.11" and python_version < "4.0" or platform_machine == "AMD64" and python_version >= "3.11" and python_version < "4.0" or platform_machine == "win32" and python_version >= "3.11" and python_version < "4.0" or platform_machine == "WIN32" and python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.60.1 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.60.1 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.2 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.25.2 ; python_version >= "3.11" and python_version < "4.0"
idna==3.6 ; python_version >= "3.11" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
mako==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.11" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.11" and python_version < "4.0"
passlib[bcrypt]==1.7.4 ; python_version >= "3.11" and python_version < "4.0"
pluggy==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.23.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==4.25.2 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.11" and python_version < "4.0"
pydantic==1.10.13 ; python_version >= "3.11" and python_version < "4.0"
pymongo==4.6.1 ; python_version >= "3.11" and python_version < "4.0"
pytest-asyncio==0.21.1 ; python_version >= "3.11" and python_version < "4.0"
pytest==7.4.3 ; python_version >= "3.11" and python_version < "4.0"
python-jose[cryptography]==3.3.0 ; python_version >= "3.11" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.11" and python_version < "4.0"
pywin32==306 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.23 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
testcontainers-core==0.0.1rc1 ; python_version >= "3.11" and python_version < "4.0"
testcontainers-mongodb==0.0.1rc1 ; python_version >= "3.11" and python_version < "4.0"
testcontainers-postgres==0.0.1rc1 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.21.1 ; python_version >= "3.11" and python_version < "4.0"
wrapt==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
